Feat: Explicitly include extremes in int generators
Change behaviour of integer generators to having 2% extra occurrences
of extreme values (minimum and maximum).

The value of this is that you have a completely random generator of
all possible u64-type values, the chance of actualy testing the
extreme values (in this case 0 and 2^64-1) is extremely small in the
100 examples actually applied to a propery. Testing with the extreme
values are important, if you want to find bugs.

In addition to to the changed behaviour described above, also add a
completely random version of int generator, that can be used where
complete randomness is important.