Refactor!: Use boxed Gen and Shrink by default
...everywhere in the Monkey Test API.

This enables simplifying the Monkey Test API by reducing the need for
extensive use of generics, especially in nested decorator-style structs
having other generators and shrinkers within.

Also, making Gen and Shrink into true trait objects, enable using
arbitrary collections of generators and shrinkers in the API. This will
come in handy when adding the possbility to mix generators and
shrinkers respectively.

Further, the number of exposed concrete implementations (structs) of
the generator and shrinker traits, can be reduced in the public API.
It should simplify the details for library user, making it more user
friendly. Potential performance loss of always using Box everyhere is
not a problem until shown otherwise.

Additional changes:
- Rename SomeIter to BoxIter, so it aligns with BoxGen and BoxShrink.
- Expose Box type aliases in API.