Feat: Catch panic as property failures in all asserts
A property can panic for some example even tough for instance asserting
equality. These panics should also be caught and treated as a property
failure.

Previously, unexpectedly panicing in non-panic-related test, monkey_test
did not even say for which example it occurred, basicly just aborting
the normal monkey_test procedure. This is now fixed and panics are
caught and treated as any other failure in the monkey_test procedure.