Fix!: Rename module 'gen' to 'gens'
In rust edition 2024, the name 'gen' is a reserved keyword, so
therefor renaming module 'gen' to 'gens.

Since the entire module is a collection of genereators the
plural 's' is a good fit and does not change the API nor
nomenclature too much. Alternative names would be something like
'example' or 'source', but with larger effects on the naming.

By renaming module to 'gens', 2024 Rust edition users of library
do not need to refer to the module with 'r#gen' but instead
just 'gens'.

The Monkey Test library it self is still kept at Rust edition
2021, just to be able to stay at the current MSRV and be usable
to as many users as possible.

For additional info see https://doc.rust-lang.org/edition-guide/rust-2024/gen-keyword.html

BREAKING CHANGE