Feat!: Add size argument to generators
BREAKING CHANGE: The trait function `Gen::examples` gets a new argument
called `size` of type (alias) `ExampleSize`. The purpose is to have
the possibility to specify the size of the examples in a generator.
This is not needed in all generators, but is needed in some generators
like generators for vectors and strings.