Fix: Make shrinking_challenge test more stable
Most often the end result is `vec![0,1]`, but some times it is the
reversed `vec![1,0]`.

This will always be the case until vector shrinking also tries to
sort (change order) of the individual elements.

The temporary solution is to "manually" sort the result before
assertion - loosening up the test somewhat.

Before fix, flakyness is reproduced in average after 64 loops. This is
based on 4 test runs with 68, 137, 19, 33 loops respectively before failing.
After fix, flakyness was not reproduced in 2300 test loops.